services/document-indexer/requirements.txt
# Document Indexer Dependencies
# Enhanced version with metadata extraction, AI services, and HTTP API

# Object Storage
minio==7.2.5

# Vector Database
qdrant-client>=1.11.0

# Document Parsing
PyMuPDF==1.24.0
pdfplumber==0.11.0
python-docx==1.1.0
markdown==3.5.2
PyYAML==6.0.1

# OCR Support (optional - for scanned PDFs)
pdf2image==1.16.3
Pillow==10.2.0

# HTTP Client
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# HTTP API Server
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-magic==0.4.27

# BM25 Search with German stemming
bm25s==0.2.0
PyStemmer==2.2.0.1

# German decompounding
compound-split>=1.0.1
